requirements.txt
flask
reportlab
python-dotenv
pillow
flask-limiter
flask-session
google-generativeai
Flask-Babel
PyPDF2
